Script 관련 함수 자동화

+++ Project/Engine/struct.h
@@ -143,7 +143,7 @@ struct tMtrlConst
 	Vec4 v4Arr[4];
 	Matrix matArr[4];
 
-	int	bTex[TEX_PARAM::END];
+	int	bTex[(UINT)TEX_PARAM::END];
 
 	int iPadding[2];
 };
